feat: Graceful shutdown upon signal
Gracefully shutdown the application upon receiving a signal.
This could be "ctrl+c" on the command line or a signal from an orchestration
system like systemd, container runtimes, etc.

This requires running the HTTP server in it's own goroutine which
slightly complicates testing.

The grace periods for handling in-transit HTTP requests and active connections
are both hard coded as constants for now.